backend/requirements.txt
Django>=5.2.9
django-cors-headers==4.3.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.45
alembic==1.17.2
